# specs/design/api-contracts.schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OnboardX API",
    "version": "1.0.0",
    "description": "Customer onboarding and KYC platform (synthetic data). Integer-only numerics. See specs/design/api-contracts.md."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "local verification mode"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/rule-sets": {
      "get": {
        "operationId": "listRuleSets",
        "summary": "All rule-set versions",
        "tags": [
          "admin"
        ],
        "x-story": "E3-S4",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetList"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createRuleSetDraft",
        "summary": "Create DRAFT copy of latest version",
        "tags": [
          "admin"
        ],
        "x-story": "E3-S4",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSet"
                }
              }
            }
          },
          "409": {
            "description": "DRAFT_EXISTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/rule-sets/{version}": {
      "get": {
        "operationId": "getRuleSet",
        "summary": "One rule-set version",
        "tags": [
          "admin"
        ],
        "x-story": "E3-S4",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Rule set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSet"
                }
              }
            }
          },
          "404": {
            "description": "Unknown version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "put": {
        "operationId": "updateRuleSet",
        "summary": "Edit a DRAFT",
        "tags": [
          "admin"
        ],
        "x-story": "E3-S4",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSet"
                }
              }
            }
          },
          "404": {
            "description": "Unknown version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "RULESET_IMMUTABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Floats or invalid types",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSetUpdate"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/rule-sets/{version}/publish": {
      "post": {
        "operationId": "publishRuleSet",
        "summary": "Publish a DRAFT",
        "tags": [
          "admin"
        ],
        "x-story": "E3-S4",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSet"
                }
              }
            }
          },
          "404": {
            "description": "Unknown version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "RULESET_IMMUTABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "RULESET_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/v1/admin/watchlist": {
      "get": {
        "operationId": "listWatchlist",
        "summary": "Watchlist entries",
        "tags": [
          "admin"
        ],
        "x-story": "E3-S4",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistList"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ]
      },
      "post": {
        "operationId": "addWatchlistEntry",
        "summary": "Add entry",
        "tags": [
          "admin"
        ],
        "x-story": "E3-S4",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistCreate"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/watchlist/{entry_id}/deactivate": {
      "post": {
        "operationId": "deactivateWatchlistEntry",
        "summary": "Deactivate entry (appends row)",
        "tags": [
          "admin"
        ],
        "x-story": "E3-S4",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntry"
                }
              }
            }
          },
          "404": {
            "description": "Unknown entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "ALREADY_DEACTIVATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNAUTHENTICATED",
          "FORBIDDEN",
          "NOT_FOUND",
          "INVALID_STATE",
          "CASE_LOCKED",
          "PROFILE_LOCKED",
          "RULESET_IMMUTABLE",
          "FILE_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "VALIDATION_ERROR",
          "UNKNOWN_CHECKLIST_ITEM",
          "MISSING_DOCUMENTS",
          "MISSING_PROFILE",
          "MISSING_PROFILE_FIELD",
          "RULESET_INVALID",
          "UNKNOWN_REASON_CODE",
          "DRAFT_EXISTS",
          "ALREADY_DEACTIVATED"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "error"
        ]
      },
...
      "RuleSet": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED"
            ]
          },
          "weights": {
            "type": "object",
            "properties": {
              "age": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100
              },
              "income_band": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100
              },
              "occupation_category": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100
              },
              "geography": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100
              }
            },
            "additionalProperties": false,
            "required": [
              "age",
              "income_band",
              "occupation_category",
              "geography"
            ]
          },
          "points": {
            "type": "object",
            "properties": {
              "age": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "label": {
                      "type": "string"
                    },
                    "min_years": {
                      "type": "integer"
                    },
                    "max_years": {
                      "type": "integer",
                      "nullable": true
                    },
                    "points": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 100
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "label",
                    "min_years",
                    "max_years",
                    "points"
                  ]
                }
              },
              "income_band": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "min_inr": {
                      "type": "integer"
                    },
                    "max_inr": {
                      "type": "integer",
                      "nullable": true
                    },
                    "points": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 100
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "min_inr",
                    "max_inr",
                    "points"
                  ]
                }
              },
              "occupation_category": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 100
                }
              },
              "geography": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 100
                }
              }
            },
            "additionalProperties": false,
            "required": [
              "age",
              "income_band",
              "occupation_category",
              "geography"
            ]
          },
          "geography_map": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "geography_default": {
            "type": "string"
          },
          "thresholds": {
            "type": "object",
            "properties": {
              "low_max": {
                "type": "integer",
                "minimum": 0,
                "maximum": 99
              },
              "medium_max": {
                "type": "integer",
                "minimum": 1,
                "maximum": 99
              }
            },
            "additionalProperties": false,
            "required": [
              "low_max",
              "medium_max"
            ]
          },
          "author": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "border_states": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[A-Z]{2}$"
            },
            "description": "State codes treated as DOMESTIC_BORDER when country_code is IN (DD-14). Seeded v1: JK, PB, AS."
          }
        },
        "additionalProperties": false,
        "required": [
          "version",
          "status",
          "weights",
          "points",
          "geography_map",
          "geography_default",
          "thresholds",
          "author",
          "created_at",
          "published_at",
          "border_states"
        ]
      },
      "RuleSetUpdate": {
        "type": "object",
        "properties": {
          "weights": {
            "type": "object",
            "properties": {
              "age": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100
              },
              "income_band": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100
              },
              "occupation_category": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100
              },
              "geography": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100
              }
            },
            "additionalProperties": false,
            "required": [
              "age",
              "income_band",
              "occupation_category",
              "geography"
            ]
          },
          "points": {
            "type": "object",
            "properties": {
              "age": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "label": {
                      "type": "string"
                    },
                    "min_years": {
                      "type": "integer"
                    },
                    "max_years": {
                      "type": "integer",
                      "nullable": true
                    },
                    "points": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 100
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "label",
                    "min_years",
                    "max_years",
                    "points"
                  ]
                }
              },
              "income_band": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "min_inr": {
                      "type": "integer"
                    },
                    "max_inr": {
                      "type": "integer",
                      "nullable": true
                    },
                    "points": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 100
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "min_inr",
                    "max_inr",
                    "points"
                  ]
                }
              },
              "occupation_category": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 100
                }
              },
              "geography": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 100
                }
              }
            },
            "additionalProperties": false,
            "required": [
              "age",
              "income_band",
              "occupation_category",
              "geography"
            ]
          },
          "geography_map": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "geography_default": {
            "type": "string"
          },
          "thresholds": {
            "type": "object",
            "properties": {
              "low_max": {
                "type": "integer",
                "minimum": 0,
                "maximum": 99
              },
              "medium_max": {
                "type": "integer",
                "minimum": 1,
                "maximum": 99
              }
            },
            "additionalProperties": false,
            "required": [
              "low_max",
              "medium_max"
            ]
          },
          "border_states": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[A-Z]{2}$"
            },
            "description": "State codes treated as DOMESTIC_BORDER when country_code is IN (DD-14). Seeded v1: JK, PB, AS."
          }
        },
        "additionalProperties": false
      },
      "RuleSetSummary": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED"
            ]
          },
          "author": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "required": [
          "version",
          "status",
          "author",
          "created_at",
          "published_at"
        ]
      },
      "RuleSetList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleSetSummary"
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "items"
        ]
      },
      "WatchlistEntry": {
        "type": "object",
        "properties": {
          "entry_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "list_type": {
            "type": "string",
            "enum": [
              "AML",
              "PEP"
            ]
          },
          "active": {
            "type": "boolean"
          },
          "added_at": {
            "type": "string",
            "format": "date-time"
          },
          "deactivated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "required": [
          "entry_id",
          "name",
          "aliases",
          "list_type",
          "active",
          "added_at",
          "deactivated_at"
        ]
      },
      "WatchlistList": {
        "type": "object",
        "properties": {
          "watchlist_version": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchlistEntry"
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "watchlist_version",
          "items"
        ]
      },
      "WatchlistCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "maxItems": 10
          },
          "list_type": {
            "type": "string",
            "enum": [
              "AML",
              "PEP"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "name",
          "list_type"
        ]
      },
...
    }
  }
}